Restructure Redux store
